Tweak Custom Toolchain Flag
I'm simplifying how the Custom Toolchain Flag works so that we can show it in the lunch menu and it will look like the other menu items.

Originally you had to put TARGET_KERNEL_CUSTOM_TOOLCHAIN := arm-eabi-4.9 but now with this simplification you can just put 4.9 and it knows that you are referring to arm-eabi-4.9.  Although this isn't a flexible as it was originally you can have multiple custom toolchains that are 4.9 by simply naming them arm-eabi-4.9-linaro and then placing 4.9-linaro for example if you want a linaro toolchain.

Signed-off-by: Chet Kener <Cl3Kener@gmail.com>

Conflicts:
	core/dumpvar.mk

Conflicts:
	core/tasks/kernel.mk